Prepared base configuration for integration testing. Refactored packages arrangement.
