+++ src/integration/java/org/mateusz/feature/UserPlayedLottoAndWonIntegrationTest.java
@@ -1,0 +1,12 @@
+package org.mateusz.feature;
+
+import org.junit.jupiter.api.Test;
+import org.mateusz.BaseIntegrationTest;
+
+public class UserPlayedLottoAndWonIntegrationTest extends BaseIntegrationTest {
+
+    @Test
+    public void should_user_win_and_system_should_generate_winners() {
+
+    }
+}
