+++ src/test/java/org/mateusz/domain/AdjustableClock.java
@@ -1,4 +1,4 @@
-package org.mateusz;
+package org.mateusz.domain;
 
 import java.time.*;
