+++ src/test/java/org/mateusz/domain/numbergenerator/WinningNumbersGeneratorTestImpl.java
@@ -1,6 +1,4 @@
-package org.mateusz.numbergenerator;
-
-import org.mateusz.domain.numbergenerator.RandomNumberGenerator;
+package org.mateusz.domain.numbergenerator;
 
 import java.util.Set;
