+++ src/test/java/org/mateusz/domain/numberreceiver/NumberReceiverFacadeTest.java
@@ -1,13 +1,10 @@
-package org.mateusz.numberreceiver;
+package org.mateusz.domain.numberreceiver;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mateusz.AdjustableClock;
+import org.mateusz.domain.AdjustableClock;
 import org.mateusz.domain.drawdate.DrawDateFacade;
 import org.mateusz.domain.drawdate.dto.DrawDateDto;
-import org.mateusz.domain.numberreceiver.DrawDateMapper;
-import org.mateusz.domain.numberreceiver.NumberReceiverConfiguration;
-import org.mateusz.domain.numberreceiver.NumberReceiverFacade;
 import org.mateusz.domain.numberreceiver.dto.NumberReceiverResponseDto;
 import org.mateusz.domain.numberreceiver.dto.TicketDto;
