+++ src/test/java/org/mateusz/domain/resultchecker/ResultCheckerFacadeTest.java
@@ -1,11 +1,10 @@
-package org.mateusz.resultchecker;
+package org.mateusz.domain.resultchecker;
 
 import org.junit.jupiter.api.Test;
 import org.mateusz.domain.numbergenerator.WinningNumbersGeneratorFacade;
 import org.mateusz.domain.numbergenerator.dto.WinningNumbersDto;
 import org.mateusz.domain.numberreceiver.NumberReceiverFacade;
 import org.mateusz.domain.numberreceiver.dto.TicketDto;
-import org.mateusz.domain.resultchecker.*;
 import org.mateusz.domain.resultchecker.dto.PlayerResultDto;
 import org.mateusz.domain.resultchecker.dto.PlayersResultsDto;
